requirement.txt
Flask
Flask-Cors

langchain
langchain-community
langchain-openai

faiss-cpu
openai

python-dotenv

requests
numpy
tqdm

# Phụ thuộc phát triển
werkzeug